+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1493](https://github.com/feldera/feldera/pull/1493))
 - SQL: support for ARRAY_PREPEND function
   ([#1496](https://github.com/feldera/feldera/pull/1496))
-
+- adapters: add an input connector for Amazon S3 (#1485)
 
 ## [0.11.0] - 2024-03-05